==> 1_aledb_mutation_df.ipynb <==
%matplotlib inline
import matplotlib.pyplot as plt
import pandas as pd
from IPython.display import display
import os, sys, itertools, csv
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from util.alemutdf import get_all_sample_mut_df, get_multi_exp_max_freq_mut_df
from util.mut import is_coding_mut
from util.metadata import get_condition_val_dict, get_condition_field_val_set
from util.genome import get_K12_pos_from_BOP27

muts_df = get_all_sample_mut_df("./data/muts/published/")
display(set(muts_df["exp"]))

l = [
    "42c_tenaillon",
    "ltee",
    "ltee_ara",
    "tee",
    "tee_ara",
    "tm_glugly",
    "epistasis2011",
    'C321',
    'C321.∆A',
    'C321.∆A.earlyfix',
    'ECNR2.1',
    'Auxotrophic_strains_hisD_gltA',
    'Auxotrophic_strains_hisD_gltB',
    'Auxotrophic_strains_hisD_pyrC',
]
exp_to_remove_l = [x.lower() for x in l]
muts_df = muts_df[~muts_df.exp.str.lower().isin(exp_to_remove_l)]
set(muts_df.exp)

## remove problem and starting strain mutations

muts_df = muts_df[~(
    ((muts_df.Gene == "gatC")
     | (muts_df.Gene == "gatC, gatC"))
    & ((muts_df.Position == "2,173,361")
       | (muts_df.Position == "2,173,363")
       | (muts_df.Position == "2,173,364"))
    & ((muts_df["Sequence Change"] == "Δ2 bp")
       | (muts_df["Sequence Change"] == "Δ1 bp"))
)]

muts_df = muts_df[~(
    (muts_df.Gene == "[crl]")
    & (muts_df.Position == "257,908")
    & (muts_df["Sequence Change"] == "Δ776 bp"))]

muts_df = muts_df[~(
    (muts_df.Gene == "insB1, insA")
    & (muts_df.Position == "1,978,503")
    & (muts_df["Sequence Change"] == "Δ776 bp"))]

muts_df = muts_df[~(
    (muts_df.Gene == "ychE, oppA")
    & (muts_df.Position == "1,299,499")
    & (muts_df["Sequence Change"] == "Δ1,199 bp"))]

muts_df = muts_df[~(
    ((muts_df.Gene == "glpR")
     | (muts_df.Gene == "glpR, glpR"))
    & (muts_df.Position == "3,560,455")
    & (muts_df["Sequence Change"] == "+G"))]

muts_df = muts_df[~(
    (muts_df.Gene == "gltP, yjcO")
    & (muts_df.Position == "4,296,381")
    & (muts_df["Sequence Change"] == "+GC"))]

muts_df = muts_df[~(
    (muts_df.Gene == "gltP, yjcO")
    & (muts_df.Position == "4,296,380")
    & (muts_df["Sequence Change"] == "+CG"))]

muts_df = muts_df[~(
    (muts_df.Gene == "insB1,insA")
    & (muts_df.Position == "1,978,503")
    & (muts_df["Sequence Change"] == "Δ776 bp"))]

muts_df = muts_df[~(
    ((muts_df.Gene == "insA/uspC") | (muts_df.Gene == "insA, uspC"))
    & (muts_df.Position == "1,979,486")
    & (muts_df["Sequence Change"] == "IS5 (+) +4 bp"))]

p = {'4,293,212',
 '4,293,403',
 '4,293,430',
 '4,296,060',
 '4,296,154',
 '4,296,189',
 '4,296,190',
 '4,296,191',
 '4,296,267',
 '4,296,268',
 '4,296,286',
 '4,296,380',
 '4,296,382'}
ch = {'+C', '+G', 'A→C', 'A→G', 'A→T', 'C→A', 'C→T', 'T→C'}
muts_df = muts_df[~(
    (muts_df.Gene == "gltP, yjcO")
    & (muts_df["Position"].isin(p))
    & (muts_df["Sequence Change"].isin(ch))
       )]

muts_df.head()

## Remove experiment specific starting strain mutations

# Filter the following mutations out if they come from specific experiments.
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="cyoB") & (muts_df["Details"]=="W190* (TGG→TAG)")))]
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="eno") & (muts_df["Details"]=="L193Q (CTG→CAG)")))]
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="chbF") & (muts_df["Details"]=="V133E (GTA→GAA)")))]
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="fumC") & (muts_df["Details"]=="G354G (GGT→GGC)")))]
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="yegE") & (muts_df["Details"]=="S683Y (TCC→TAC)")))]
muts_df = muts_df[(~((muts_df.exp=="SER") & (muts_df["Gene"]=="rcsC") & (muts_df["Details"]=="L594F (TTG→TTC)")))]


muts_df = muts_df[(~((muts_df.exp=="GYD") & (muts_df["Gene"]=="fumC") & (muts_df["Details"]=="G354G (GGT→GGC)")))]
muts_df = muts_df[(~((muts_df.exp=="GYD") & (muts_df["Gene"]=="yegE") & (muts_df["Details"]=="S683Y (TCC→TAC)")))]
muts_df = muts_df[(~((muts_df.exp=="GYD") & (muts_df["Gene"]=="rcsC") & (muts_df["Details"]=="L594F (TTG→TTC)")))]


cck_exp_list = [
    "Central_carbon_knockout_tpiA",
    "Central_carbon_knockout_PTS",
    "Central_carbon_knockout_pgi",
    "Central_carbon_knockout_sdh",
    "Central_carbon_knockout_gnd",
    "Central_carbon_knockout_glucose_evolution"
]

muts_df = muts_df[~(
    (muts_df.exp.isin(cck_exp_list))
    & (
        (muts_df["Gene"]=="wcaA")
        | (muts_df["Gene"]=="corA")
    )
)]  # The details of the mutations to filter out should probably include the sequence change.

# Since the majority of our comparisons have to do with endpoints, just look at endpoints.
# If not limiting to endpoint, the mutations used in associated mutations to conditions will need filtering.
# This occurs in a later NB of the pipeline.
muts_df = get_multi_exp_max_freq_mut_df(muts_df, endpoint_flask_only=True)
muts_df.head()

muts_df.Position = muts_df.Position.apply(lambda x: int(str(x).replace(",","")))

# ensure all mutations are using NC_000913 positions.
genome_pos_transform_exp_l = ["ERS",
                              'Central_carbon_knockout_PTS',
                              'Central_carbon_knockout_glucose_evolution',
                              'Central_carbon_knockout_gnd',
                              'Central_carbon_knockout_pgi',
                              'Central_carbon_knockout_sdh',
                              'Central_carbon_knockout_tpiA',]
muts_df.Position = muts_df.apply(lambda row: get_K12_pos_from_BOP27(row.Position) if row.exp in genome_pos_transform_exp_l else row.Position, axis=1)

muts_df["coding"] = muts_df.Details.apply(is_coding_mut)
muts_df["Gene"] = muts_df.apply(lambda row: row["Gene"].replace(", ","/") if not row["coding"] else row["Gene"], axis=1)

muts_df.head()

print(len(muts_df))
muts_df = muts_df[muts_df.presence >= 0.5]
print(len(muts_df))

# This work is also currently duplicated NB4 
muts_df["Gene"] = muts_df["Gene"].apply(lambda a: "rph" if a == "[rph], [rph]" else a)
muts_df["Gene"] = muts_df["Gene"].apply(lambda a: "rph" if a == "[rph],[rph]" else a)
muts_df.head()

# removing unnecessary column
muts_df = muts_df.drop(columns=["Mut ID"])

# Amount of ALEs before removal
print(len(muts_df.apply(lambda r: r.exp + " " + str(r.ale), axis=1).unique()))

# removing hypermutators since they will affect associations

muts_df = muts_df[
    ~(
        (muts_df.exp == "GLU")
        & ((muts_df.ale == 5) | (muts_df.ale == 7))
    )
]

muts_df = muts_df[
    ~(
        (muts_df.exp == "TOL_2,3-butanediol")
        & (muts_df.ale == 3)
    )
]

muts_df = muts_df[
    ~(
        (muts_df.exp == "TOL_adipic_acid")
        & (muts_df.ale == 5)
    )
]

muts_df = muts_df[
    ~(
        (muts_df.exp == "TOL_hexamethylenediamine")
        & (muts_df.ale == 6)
    )
]

muts_df = muts_df[
    ~(
        (muts_df.exp == "TOL_octanoic_acid")
        & (muts_df.ale == 5)
    )
]

muts_df = muts_df[
    ~(
        (muts_df.exp == "TOL_propanediol")
        & ((muts_df.ale == 1) | (muts_df.ale == 2) | (muts_df.ale == 3) | (muts_df.ale == 4) | (muts_df.ale == 5) | (muts_df.ale == 7) | (muts_df.ale == 8))
    )
]

print(len(muts_df.apply(lambda r: r.exp + " " + str(r.ale), axis=1).unique()))

# removing samples not incuded in GYD MS for AVA MS

muts_df = muts_df[
    ~(
        (muts_df.exp == "GYD")
        & ((muts_df.ale == 35) | (muts_df.ale == 34))
    )
]
print(len(muts_df.apply(lambda r: r.exp + " " + str(r.ale), axis=1).unique()))

muts_df.to_pickle("./data/1_df.pkl")

display(len(muts_df.exp.unique()))

for e in muts_df.exp.unique():
    print(e)

muts_df["exp ALE"] = muts_df.apply(lambda r: r.exp + ' ' + str(r.ale), axis=1)
display(len(muts_df["exp ALE"].unique()), muts_df["exp ALE"].unique())

len(muts_df)
