==> 2_2_annot_genomic_feats.ipynb <==
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.ticker import MaxNLocator
import pandas as pd
import seaborn as sns
from IPython.display import display
import os, sys, itertools, csv
module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from util.metadata import get_condition_val_dict, get_condition_field_val_set
from util.gene import get_coding_genetic_target_len_d, get_intergenic_len_d
from util.genome import get_feature_hit_set, is_overlap, get_promoter_range_from_RegulonDB_df_row
pd.options.display.max_columns = 100

all_muts_df = pd.read_pickle("./data/2_df.pkl")
display(all_muts_df.shape, all_muts_df.head())

# Add component ranges
Need to find genetic features first to serve as defaults for when genomic features can't be identified.

gene_df = pd.read_pickle("./data/gene_df.pkl")
gene_df["range"] = gene_df.apply(lambda row: (row["GENE_POSLEFT"], row["GENE_POSRIGHT"]), axis=1)
TU_df = pd.read_pickle("./data/TU_df.pkl")
operon_df = pd.read_pickle("./data/operon_df.pkl")

gene_df.head()

tu_objects_df = pd.read_pickle("./data/TU_objects_df.pkl")

def get_operon_id(TU_ID, TU_df):
    operon_id = ""
    df = TU_df[TU_df["TRANSCRIPTION_UNIT_ID"]==TU_ID]
    if len(df) > 0:
        operon_id = list(df["OPERON_ID"])[0]
    return operon_id


def get_operon_name_from_operon_ID(operon_ID, operon_df):
    operon_name = "unknown"
    df = operon_df[operon_df["OPERON_ID"]==operon_ID]
    if len(df) > 0:
        operon_name = list(df["OPERON_NAME"])[0]
    return operon_name


tu_objects_df["OPERON_ID"] = tu_objects_df["TRANSCRIPTION_UNIT_ID"].apply(get_operon_id, args=[TU_df])
tu_objects_df["OPERON_NAME"] = tu_objects_df["OPERON_ID"].apply(get_operon_name_from_operon_ID, args=[operon_df])

tu_objects_df[:10]

# tfbs_df = pd.read_csv("./data/RegulonDBwebsite10/BindingSiteSet.txt", sep="\t", comment='#', header=None)
tfbs_df = pd.read_pickle("./data/TFBS_df.pkl")


def get_TFBS_range(tfbs_df_row):
    r = ()
    if not pd.isna(tfbs_df_row[3]) and not pd.isna(tfbs_df_row[4]):
        r = (int(tfbs_df_row[3]), int(tfbs_df_row[4])) 
    return r


tfbs_df["range"] = tfbs_df.apply(get_TFBS_range, axis=1)
display(tfbs_df.shape, tfbs_df.head())

# promoter_df = pd.read_csv("./data/RegulonDB10/promoter.txt", sep="\t", comment='#', header=None)
promoter_df = pd.read_pickle("./data/promoter_df.pkl")
promoter_df["range"] = promoter_df.apply(get_promoter_range_from_RegulonDB_df_row, axis=1)
promoter_df.head()


# def get_TSS_range_from_row(promoter_df_row):
#     r = ()
#     promoter_seq_str = promoter_df_row[9]
#     if not pd.isna(promoter_seq_str):
#         TSS_genome_pos = int(promoter_df_row[3])
#         r = (TSS_genome_pos, TSS_genome_pos)
#     return r

# promoter_df["TSS range"] = promoter_df.apply(get_TSS_range_from_row, axis=1)
# promoter_df.head()

att_term_df = pd.read_pickle("./data/att_term_df.pkl")
att_term_df.head()

# terminator_df = pd.read_csv("./data/RegulonDB10/terminator.txt", sep="\t", comment='#', header=None)
terminator_df = pd.read_pickle("./data/term_df.pkl")
terminator_df["range"] = terminator_df.apply(lambda row: (row[2], row[3]), axis=1)
terminator_df.head()

RBS_df = pd.read_pickle("./data/RBS_df.pkl")
RBS_df.head()

from util.mut import is_genetic_mut
all_muts_df["genetic"] = all_muts_df["Details"].apply(is_genetic_mut)
all_muts_df.head()

# Build component annotation

NONCODING_FEATURE_COLUMNS = [
    "TFBS",
    "promoter",
#     "TSS",
    "RBS",
    "terminator",
    "attenuator terminator"
]


CODING_FEATURE_COLUMNS = ["gene RegulonDB ID"]


FEATURE_COLUMNS = NONCODING_FEATURE_COLUMNS + CODING_FEATURE_COLUMNS


def get_gene_name(RegulonDB_ID, debug=False):
    if debug:
        display(RegulonDB_ID, gene_df[gene_df["GENE_ID"] == RegulonDB_ID])
    return gene_df[gene_df["GENE_ID"] == RegulonDB_ID].iloc[0]["GENE_NAME"]  #Simply returning the first gene name in the second column that has the RegulonDB ID in the first column.


def get_feature_range(feat_id, feat_type):
    r = ()
    # default parameters
    feat_id_col = 0
    range_col = "range"
    if feat_type == "gene RegulonDB ID":
        feat_id_col = "GENE_ID"
        feat_type_df = gene_df
    elif feat_type == "TFBS":
        feat_id_col = 2
        feat_type_df = tfbs_df
    elif feat_type == "promoter":
        feat_type_df = promoter_df
#     elif feat_type == "TSS":
#         feat_type_df = promoter_df
#         range_col = "TSS range"
    elif feat_type == "RBS":
        feat_id_col = "SHINE_DALGARNO_ID"
        feat_type_df = RBS_df
    elif feat_type == "terminator":
        feat_type_df = terminator_df
    elif feat_type == "attenuator terminator":
        feat_id_col = "RegulonDB ID"
        feat_type_df = att_term_df
    
    df = feat_type_df[feat_type_df[feat_id_col] == feat_id]
    if not df.empty:
        r = df.iloc[0][range_col]  # Assuming all rows describe the same range.
    l = [int(i) for i in r]
    return tuple(l)


def _get_operon_name(TU_object_ID):
    operon_name = "unknown"
    df = tu_objects_df[tu_objects_df["TU_OBJECT_ID"] == TU_object_ID]
    if len(df) > 0:
        # non-coding features can be associated to more than one TU.
        operon_name_set = set([])
        for _, row in df.iterrows():
            operon_name_set.add(row["OPERON_NAME"])
        operon_name = ' '.join(operon_name_set)
    return operon_name


def get_genomic_feature_name(feat_id, feat_type):
    name = ""
    if feat_type == "gene RegulonDB ID":
        name = get_gene_name(feat_id)
    elif feat_type == "TFBS":  # TFBS can be assoicated to more than one TU.
        for _, row in tfbs_df[tfbs_df[2]==feat_id].iterrows():
            name += row[7] + " "
        name += "TFBS"
    elif feat_type == "promoter":
        name = promoter_df[promoter_df[0]==feat_id].iloc[0][1]
#     elif feat_type == "TSS":
#         name = promoter_df[promoter_df[0]==feat_id].iloc[0][1] + " TSS"
    elif feat_type == "RBS":
        gene_id = RBS_df[RBS_df["SHINE_DALGARNO_ID"]==feat_id].iloc[0][1]
        name = get_gene_name(gene_id) + " RBS"
    elif feat_type == "terminator":
        name = _get_operon_name(feat_id) + " " + "terminator"
    elif feat_type == "attenuator terminator":
        name = _get_operon_name(feat_id) + " " + "attenuator terminator"
    return name


def get_genomic_features(mut_row, ignore_RegDB_feats=False):
    annots = []
    if not ignore_RegDB_feats:
        for feat_col in FEATURE_COLUMNS:
            for feat_id in mut_row[feat_col]:
                d = {
                    "name": get_genomic_feature_name(feat_id, feat_col),
                    "RegulonDB ID": feat_id,
                    "range": get_feature_range(feat_id, feat_col),
                    "genetic": False if feat_col in NONCODING_FEATURE_COLUMNS else True,  # could be pseudogene which is non-coding though still genetic.
                    "feature type": "gene" if feat_col == "gene RegulonDB ID" else feat_col,
                    "operon": _get_operon_name(feat_id)
                }
                annots.append(d)
                
    if len(annots) == 0:
        annots = mut_row["genetic features"]  # Just use what's in the genetic features col.
        for a in annots:
            a["genetic"] = False if '/' in a["RegulonDB ID"] else True
            # Currently not getting operon here for integenic. Rather only getting it within the operon NB, where it's used for intergenic.
        
    return annots

all_muts_df["genomic features"] = all_muts_df.apply(lambda r: get_genomic_features(r), axis=1)
all_muts_df.head()

# Define if features are found to overlap
for _, r in all_muts_df.iterrows():
    if len(r["genomic features"]) > 1:
        for i in range(0, len(r["genomic features"])):
            feat_d1 = r["genomic features"][i]
            for j in range(i+1, len(r["genomic features"])):
                feat_d2 = r["genomic features"][j]                
                if is_overlap(feat_d1["range"], feat_d2["range"]):
                    if "overlap" not in feat_d1.keys(): feat_d1["overlap"] = set()
                    feat_d1["overlap"].add(feat_d2["name"])
                    if "overlap" not in feat_d2.keys(): feat_d2["overlap"] = set()
                    feat_d2["overlap"].add(feat_d1["name"])

# Remove terminators that overlap with attenuator-terminators
# The following requires all indexes to be unique
all_muts_df = all_muts_df.reset_index()

for i, r in all_muts_df.iterrows():
    if len(r["genomic features"]) > 1:
        idxs_to_remove = []
        for k in range(0, len(r["genomic features"])):
            # finding a mutated attenuator-terminator that also overlaps with a mutated terminator from the same mutation.
            feat_d = r["genomic features"][k]
            if "overlap" in feat_d.keys() \
            and feat_d["feature type"] == "terminator" \
            and feat_d["operon"]+' '+"attenuator terminator" in feat_d["overlap"]:
                idxs_to_remove.append(k)
        if len(idxs_to_remove) > 0:
            # removing the overlapping attenuator terminator
            feats = [r["genomic features"][j] for j in range(0, len(r["genomic features"])) if j not in idxs_to_remove]
            all_muts_df.at[i, "genomic features"]=feats

# # test code to check if any features mutated within ALEdb are completely overlapped by a gene.


# gene_df = pd.read_csv(
#     "./data/RegulonDB10/gene.txt", sep="\t", comment='#', header=None)
# gene_df.columns = [
#     "GENE_ID",
#     "GENE_NAME",
#     "GENE_POSLEFT",
#     "GENE_POSRIGHT",
#     "GENE_STRAND",
#     "GENE_SEQUENCE",
#     "GC_CONTENT",
#     "CRI_SCORE",
#     "GENE_NOTE",
#     "GENE_INTERNAL_COMMENT",
#     "KEY_ID_ORG",
#     "GENE_TYPE"
# ]
# display(len(gene_df), gene_df.head())


# for i, mut in all_muts_df.iterrows():
#     for geno_feat_d in mut["genomic features"]:
#         if not geno_feat_d["genetic"]:
#             for j, gene in gene_df.iterrows():
#                 if not pd.isna(gene.GENE_POSLEFT) and not pd.isna(gene.GENE_POSRIGHT):
#                     if (geno_feat_d["range"][0] >= gene.GENE_POSLEFT) and (geno_feat_d["range"][1] <=gene.GENE_POSRIGHT):
#                         display(geno_feat_d, gene, "##########################################")

all_muts_df.to_pickle("./data/2_2_df.pkl")
