==> 2_3_annot_operons.ipynb <==
%matplotlib inline
import numpy as np
import matplotlib.pyplot as plt
from matplotlib.ticker import MaxNLocator
import pandas as pd
import seaborn as sns
from IPython.display import display
import os, sys, itertools, csv
pd.options.display.max_columns = 100

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from util.genome import get_feature_hit_set, is_overlap, NON_K12_EXP_L
from util.operon import get_operon_name_set, get_operon_ID_set
from util.feature import get_feat_d_from_ID

all_muts_df = pd.read_pickle("./data/2_2_1_df.pkl")
display(all_muts_df.shape, all_muts_df.head())

operon_df = pd.read_pickle("./data/operon_df.pkl")
RBS_df = pd.read_pickle("./data/RBS_df.pkl")
TU_df = pd.read_pickle("./data/TU_df.pkl")
TU_objects_df = pd.read_pickle("./data/TU_objects_df.pkl")
TU_objects_df["TU range"] = TU_objects_df.apply(
    lambda row: (row["TU_POSLEFT"], row["TU_POSRIGHT"]), axis=1)

def get_operons(mut_row, operon_df):
    annots = []
    if mut_row.exp not in NON_K12_EXP_L:
        # 0) for each geno feat mutated by this individual mutation
        for geno_feat_d in mut_row["genomic features"]:
#             display(geno_feat_d)
            if '/' not in geno_feat_d["RegulonDB ID"]:
                # 1) get operon for geno feat using explicit relationship from RegDB.
                regdbid_to_use_for_op = geno_feat_d["RegulonDB ID"]
                if geno_feat_d["feature type"] == "RBS":
                    regdbid_to_use_for_op = RBS_df[RBS_df["SHINE_DALGARNO_ID"]==geno_feat_d["RegulonDB ID"]]["GENE_ID"].iloc[0]
                    
                op_ID_set = get_operon_ID_set(regdbid_to_use_for_op, TU_objects_df, TU_df, operon_df)
                # 2) annot each operon that feature is involved in individually within JSON.
                for op_ID in op_ID_set:
                    op_df = operon_df[operon_df["OPERON_ID"] == op_ID]
                    for _, op in op_df.iterrows():
                        d = {
                            "name": op["OPERON_NAME"],
                            "RegulonDB ID": op["OPERON_ID"],
                            "range": (op["REGULATIONPOSLEFT"], op["REGULATIONPOSRIGHT"])}
                        annots.append(d)

            # This will catch whether no features exist in the feat_ID_set
            # or if those features are just integenic regions.
            # Get operon genetic feature link dict through overlapping mutation and operon ranges.
            # All genetic features are annotated within both the "genetic features" and "genetic feature links" column.
            # If no integenic genomic feature is hit by a mutation, use the genetic feature range in an overlap analysis with operons.

            else:
                operon_df["mutation hit"] = operon_df["range"].apply(is_overlap, args=[geno_feat_d["range"]])
                op_mut_hit_df = operon_df[operon_df["mutation hit"]]
                for _, op in op_mut_hit_df.iterrows():
                    d = {
                        "name": op["OPERON_NAME"],
                        "RegulonDB ID": op["OPERON_ID"],
                        "range": (op["REGULATIONPOSLEFT"], op["REGULATIONPOSRIGHT"])}
                    annots.append(d)
                operon_df = operon_df.drop(columns=["mutation hit"])  # Removing this column to clean up for next iteration.

        return annots
    

all_muts_df["operons"] = all_muts_df.apply(lambda r: get_operons(r, operon_df), axis=1)
all_muts_df.head()

# TODO: combine this with the get_operons() function above

def _get_genetic_target_d_from_ID(gen_targ_ID, mut_row):
    gen_targ_d = dict()
    for d in mut_row["genetic features"]:
        if d["RegulonDB ID"] == gen_targ_ID:
            gen_targ_d = d
            break
    return gen_targ_d


def get_operon_genetic_link_d(mut_row, operon_df):
    op_gen_targ_d = dict()
    if mut_row.exp not in NON_K12_EXP_L:
        # 0) for each genetic and genomic feat mutated by this individual mutation
        for gen_targ_ID, feat_ID_set in mut_row["genetic feature links"].items():  # Assuming all genetic features in mutation are defined here
            # 1) for each genomic feat mutated by this mut.
            for feat_ID in feat_ID_set:
                # 2) get the dict for this genome feat because we want it's regdb ID.
                feat_d = get_feat_d_from_ID(feat_ID, mut_row)
#                 link_exists = False
                
                if '/' not in feat_d["RegulonDB ID"]:
                    regdbid_to_use_for_op = feat_d["RegulonDB ID"]
                    if feat_d["feature type"] == "RBS":
                        regdbid_to_use_for_op = RBS_df[RBS_df["SHINE_DALGARNO_ID"]==feat_d["RegulonDB ID"]]["GENE_ID"].iloc[0]
                    
                    # 3) get operon for geno feat using explicit relationship from RegDB.
                    op_ID_set = get_operon_ID_set(regdbid_to_use_for_op, TU_objects_df, TU_df, operon_df)
#                     if len(op_ID_set) > 0: link_exists = True
                    for op_ID in op_ID_set:
                    # 4) annot the GENETIC FEAT that this 
                        if op_ID not in op_gen_targ_d.keys():
                            op_gen_targ_d[op_ID] = []
                        op_gen_targ_d[op_ID].append(gen_targ_ID)

                else:
                    operon_df["mutation hit"] = operon_df["range"].apply(is_overlap, args=[feat_d["range"]])
                    op_mut_hit_df = operon_df[operon_df["mutation hit"]]
#                     if len(op_mut_hit_df) > 0: link_exists = True
                    for _, op in op_mut_hit_df.iterrows():
                        if op["OPERON_ID"] not in op_gen_targ_d.keys():
                            op_gen_targ_d[op["OPERON_ID"]] = []
                        op_gen_targ_d[op["OPERON_ID"]].append(gen_targ_ID)
                    # cleanup
                    if "mutation hit" in operon_df.columns.values:
                        operon_df = operon_df.drop(columns=["mutation hit"])
                        
#                 if link_exists == False:
#                     if "unknown" not in op_gen_targ_d.keys(): op_gen_targ_d["unknown"] = []
#                     op_gen_targ_d["unknown"].append(gen_targ_ID)
                
        return op_gen_targ_d
    

all_muts_df["operon links"] = all_muts_df.apply(lambda r: get_operon_genetic_link_d(r, operon_df), axis=1)
all_muts_df.head()

all_muts_df.to_pickle("./data/2_3_df.pkl")
