==> 5_filter_muts.ipynb <==
import pandas as pd
pd.options.display.max_columns = 100

all_muts_df = pd.read_pickle("./data/4_7_df.pkl")
all_muts_df.shape

all_muts_df.to_csv("./data/supp/all_muts_annots.csv")

all_muts_df.exp.unique()

muts_df = all_muts_df[all_muts_df["exp"] == "GYD"]

# # This is for filtering out mutations with genomic features statistically associated to specific conditions
# muts_df = pd.DataFrame()
# for _, mut in all_muts_df.iterrows():
#     for feat in mut["genomic features"]:
        
#         if ('M9' in feat['significantly associated conditions']) \
#             & ('511145' in feat['significantly associated conditions']):

#             muts_df = muts_df.append(mut)

#             # need to break since a mutation could have more than one feature
#             # which are associated with a condition,
#             # therefore adding a duplicate entry.
#             break

muts_df.to_pickle("./data/5_df.pkl")
display(muts_df.shape,muts_df.head())
