==> 11_regulon_convergence.ipynb <==
import os, sys
import pandas as pd

module_path = os.path.abspath(os.path.join('..'))
if module_path not in sys.path:
    sys.path.append(module_path)
from util.feature import get_feat_d_from_ID
from util.genome import is_overlap
from util.params import ENRICH_ALPHA, MULTI_HYP_CORR_METHOD
pd.options.display.max_columns = 100

exp_muts_df = pd.read_pickle("./data/10_df.pkl")
display(exp_muts_df.shape, exp_muts_df.head())

## significance

# a bit behind annotations currently used in AVA MS.
# Still uses same approach to counted, just not all data-types have been consolidated into the regulators feat col


comp_mut_df = pd.DataFrame(columns=["observed mutation count"])
for i, r in exp_muts_df.iterrows():
    for f, links in r["regulator links"].items():
        if f not in comp_mut_df.index:
            df = pd.DataFrame({"observed mutation count": len(links)}, index=[f])
            comp_mut_df = comp_mut_df.append(df)
        else:
            comp_mut_df.loc[f, "observed mutation count"] += len(links)
            
comp_mut_df

for _, r in exp_muts_df.iterrows():
    for feat_d in r["regulators"]:
        feat_d["mutation set count"] = comp_mut_df.loc[feat_d["name"]]["observed mutation count"]

reg_len_df = pd.read_pickle("./data/regulon_len_df.pkl")
reg_len_df.head()

import numpy as np


comp_mut_df["length"] = comp_mut_df.apply(lambda r: reg_len_df.loc[r.index, "regulon length"])
total_seq_len = np.sum(comp_mut_df["length"])
comp_mut_df["proportion"] = comp_mut_df["length"].apply(lambda comp_len: comp_len/total_seq_len)
comp_mut_df.head()

from collections import Counter

comp_mut_df['equal or larger count'] = 0
component_l = comp_mut_df.index.tolist()
proportion_l = comp_mut_df["proportion"].tolist()
num_muts = np.sum(comp_mut_df["observed mutation count"])

# slides from UW speak about the relationship of permutations and precision.
# should read through them to get a sense of what is the least iterations that should execute.

num_itr = 10000
for i in range(0, num_itr):
    random_mut_comp_a = np.random.choice(
        a=component_l,
        size=int(num_muts),
        p=proportion_l,
        replace=True,
    )
    random_mut_comp_count_d = Counter(random_mut_comp_a)
    for component in random_mut_comp_count_d.keys():
        if random_mut_comp_count_d[component] >= comp_mut_df.loc[component, 'observed mutation count']:
            comp_mut_df.loc[component, 'equal or larger count'] += 1

comp_mut_df["p value"] = comp_mut_df["equal or larger count"].apply(
    lambda count: count/num_itr)

comp_mut_df["significant"] = comp_mut_df["p value"].apply(lambda p_val: True if p_val < ENRICH_ALPHA else False)

display(comp_mut_df.head())

from statsmodels.stats import multitest
pvals = comp_mut_df["p value"]
rejects, pvals_corrected, alphacSidak, alphacBonf = multitest.multipletests(
    pvals=pvals,
    alpha=ENRICH_ALPHA,
    method=MULTI_HYP_CORR_METHOD)

pd.options.display.max_rows = 500
comp_mut_df["corrected p value"] = pvals_corrected
comp_mut_df["corrected significance"] = rejects
comp_mut_df.head()

# # not wanting to consider mutation events with only 1 features as being significant
comp_mut_df["final significance"] = comp_mut_df.apply(lambda r: r["corrected significance"] if r["observed mutation count"] > 1 else False, axis=1)
# comp_mut_df.head()

for i, r in exp_muts_df.iterrows():
    for d in r["regulators"]:
        if d["name"] in comp_mut_df.index:
            d["significantly enriched"] = comp_mut_df.loc[d["name"]]["final significance"]
#             d["significant"] = comp_mut_df.loc[d["name"]]["final significance"]

assert(len(exp_muts_df.exp.unique())==1)
exp = exp_muts_df.exp.unique()[0]
comp_mut_df.to_csv("./data/supp/permutation_test_results_per_experiment/"+exp+"_mut_reg_signif.csv")

exp_muts_df.to_pickle("./data/11_df.pkl")

# import seaborn as sns
# import matplotlib
# import matplotlib.pyplot as plt
# %matplotlib inline
# plt.rcParams["figure.dpi"] = 100
# sns.set_context("talk")
# sns.set_style("ticks")


# def _is_signif(name, regulons):
#     is_signif = False
#     for d in regulons:
#         if d["name"] == name:
#             is_signif = d["significant"]
#     return is_signif


# reg_cnt_d = dict()
# for _, r in exp_muts_df.iterrows():
#     reg_link_d = r["regulator links"]
#     for reg, hits in reg_link_d.items():
#         t = reg        
#         if _is_signif(reg, r["regulators"]):
#             t = '*' + t
#         if t not in reg_cnt_d.keys(): reg_cnt_d[t] = 0
#         reg_cnt_d[t] += len(hits)
# df = pd.DataFrame.from_dict(reg_cnt_d, orient='index', columns=["mutated feature count"])
# df = df.sort_values("mutated feature count")
# df = df[-20:]
# ax = df.plot.barh(y='mutated feature count',
# #                   figsize=(10,25),
#                   figsize=(5,10),
#                   color="#4C72B0",
#                   width=1
#                  )
# for i, v in enumerate(df["mutated feature count"]):
#     ax.text(v, i - 0.2, str(v))
# sns.despine(ax=ax, top=True, right=True)
# ax.set_title("Top 20 Regulon mutated feature count")
# ax.get_legend().remove()

# reg_cnt_d = dict()
# for _, r in exp_muts_df.iterrows():
#     reg_link_d = r["regulator links"]
#     for reg, hits in reg_link_d.items():
#         t = reg
#         if _is_signif(reg, r["regulators"]):
#             t = '*' + t
#             if t not in reg_cnt_d.keys(): reg_cnt_d[t] = 0
#             reg_cnt_d[t] += len(hits)

# if len(reg_cnt_d) > 0:
#     df = pd.DataFrame.from_dict(reg_cnt_d, orient='index', columns=["mutated feature count"])
#     df = df.sort_values("mutated feature count")
#     ax = df.plot.barh(y='mutated feature count',
#     #                   figsize=(10,25),
#                       figsize=(5,10),
#                       color="#4C72B0",
#                       width=1
#                      )
#     for i, v in enumerate(df["mutated feature count"]):
#         ax.text(v, i - 0.2, str(v))
#     sns.despine(ax=ax, top=True, right=True)
#     ax.set_title("Significant regulon mutated feature count")
#     ax.get_legend().remove()
